Make TrustedRoots a verification option
Also export TestRawReport.

When using this library for go-tpm-tools, I found that these changes
make testing the integration much easier.

The noCrossCheckTestOnly option was clunky, so I replaced it with a more
common idiom of passing in the trusted root certificates. The default
behavior is still the same: no specific trusted root falls back on the
embedded certificates.

This also swaps around Ark and Ask validation orders since that's the
easiest way to make the Ark validation logic tests reach past the
cross-check tests for the Ask.

Signed-off-by: Dionna Glaze <dionnaglaze@google.com>