Fix Linux ABI
The ioctl numbering was incorrect and led to ENOTTY. The request size
is 32 bytes without packing, and the dirshift was incorrect.

The *Safe types and their "ABI equivalents" didn't work in a real
hardware test. We instead remove the Safe suffix and create an
ABI-specific conversion interface for correct interaction with the
Golang unsafe pointer type.

The original type for a report response did not include the 32 byte
header, so that is added as well.

Tested in a SEV-SNP guest.

Signed-off-by: Dionna Glaze <dionnaglaze@google.com>