Add a report validation library
In addition to checking the report signature with the verify library, we
also want to have basic logic for checking expected values of a report
that aren't its signature.

I expect the validation library to be more of a library template than
used directly by most users, since there may be more specific validation
logic based on use case and problem domain.

Signed-off-by: Dionna Glaze <dionnaglaze@google.com>