Add KDS cert fetching logic
An attestation report carries enough information with it (apart from
which chip product it's from (Issue #11)) that the report's certificate
chain can be reconstructed from the report's contents and AMD's Key
Distribution Service (KDS).

The attestation.bin test data is an attestation report generated on a
test SEV-SNP machine. The vcek.testcer is updated to be the certificate
for that machine at the current TCB.

Tested: The committed test is offline, but the test passes without a mock
Getter when connected to the internet.

Signed-off-by: Dionna Glaze <dionnaglaze@google.com>