Default download missing certs in verify
For hosts that don't yet provide a cache of the VCEK's certificate chain
for GetExtendedReport, the function will return an empty certificate
chain and log warnings. If the client doesn't catch this, then it's up
to the verifier to fetch the required certificates. Make fetching an
opt-out feature of verify.SnpAttestation.

Signed-off-by: Dionna Glaze<dionnaglaze@google.com>