Recognize retry error from VMM
It's possible for the host to throttle the guest request calls since
they synchronously use a shared resource: the secure processor.

If throttled, the VMM returns its result in the upper 32 bits of
EXIT_INFO_2 in the GHCB. /dev/sev-guest returns the entire EXIT_INFO_2
as fw_err, (despite only the lower 32 bits meaning firmware errors).
We check for the EBUSY error code in the upper 32 bits and use that
information to try again with exponential backoff rather than error.

We call the EBUSY value checking legacy here because we expect to get
an upstream patch into /dev/sev-guest to directly return -EAGAIN when it
detects that it's been throttled.

Signed-off-by: Dionna Glaze <dionnaglaze@google.com>