Add package to select a fake/real device in tests
This follows the pattern found in go-tpm-tools for getting a real TPM
device or a virtual one based on whether the device path was passed into
the test.

To allow its use in the verify package, the AMDRootCerts type had to be
split into its own package. Similarly in order for the client test to
not have a circular dependency, the test client code had to be separate
from the other testing code.

Signed-off-by: Dionna Glaze <dionnaglaze@google.com>