Add hardware testing capability to client_test
Like testclient, but without a circular dependency, the client_test will
open a real device if --sev_guest_device_path is not "default". To avoid
a dependence on cached certificates or a KDS connection, when a real
device is used, client_test will not check the report's signature.

Some fields of the report can't be statically determined for the test
case data, so we clear those out to pass comparison tests.

The guest policy may be different than 0xa0000 (debug true) when run on
real hardware, so the expected policy is a flag to the test.

Signed-off-by: Dionna Glaze <dionnaglaze@google.com>