Fix SevFirmwareErr, SevEsErr implementations and clean up error conditionals
An error type should be a pointer to an object that implements Error.
The indirection makes direct error comparison fail, so we switch to
the existing pattern of wantErr as a string with a Contains test.

The testing library implementation can't be used in kds_test.go due to a
cyclic dependency that would induce, but it's a mild inconvenince.

Signed-off-by: Dionna Glaze <dionnaglaze@google.com>