Refactor root x509 certs into struct
The product certificates of a root should be paired together for better
product -> cert lookups when necessary. The value type should be a pair
of the ASK and ARK.

Since AMDRootCerts is a main part of the API, this is a
backwards-incompatible change. Best to get these out of the way before
1.0.

Signed-off-by: Dionna Glaze <dionnaglaze@google.com>