Add a workaround to Linux 6.1's sev-guest lockout
The IV reuse fix 47894e0fa6a5 works against the host's throttling by
making the sev-guest device useless when throttled. The x86/urgent fix
72f7754dcf31 that merged to tip on March 16 still hasn't made it to
some distros, so this workaround is still needed temporarily.

Issue #40 is to revert this change at the appropriate time. The same
fix for #40 includes the fix for #5.

Signed-off-by: Dionna Glaze <dionnaglaze@google.com>