Update attestation_report 'flags' bitmask
Previously the bitmask only allowed bit 0 to be non-zero.
However, with the introduction of VLEKs, the SNP ABI spec
in revision 1.54 specifies that bit 0 to 4 may be non-zero.
This does not add VLEK support but allows reading an
extended guest report that is signed using a VLEK.

Also add the VlekGUID as per AMD GHCP standard revision 2.02.