Fix Issue#51 with new TCB validation
Compare TCB versions component-wise with a new TCBPartsLE function that
returns true iff all security patch levels (SPLs) of the TCB on the left are
less than or equal to (LE) the corresponding SPLs on the right.

With this change comes a policy difference, not just a bug fix. To allow
for fleet-wide reported TCBs to be lower than an individual node's
committed TCB, I've relaxed some checks. The true floor of the expected
TCB is only going to be the provided MinimumTCB. Internal relationships
between TCB values are subject to change as we learn more how we're
going to manage policies of VMs that have migrated with a migration
agent.

Signed-off-by: Dionna Glaze <dionnaglaze@google.com>