Improve product handling in attestations
Collection of the product information uses the Device interface to allow
for testing on non-SEV-SNP machines. The LinuxDevice will use CPUID
information.

Verification now uses product information from either the attestation or
the verification options. The option value takes precedence. When both
are missing, then the library defaults to the first version that was
available for use: "Milan-B0".

This extends the support of products to Genoa, but does not yet include
the embedded AMD root and intermediate certificates for Genoa.

Fixes Issue#11 and Issue#64.

Signed-off-by: Dionna Glaze <dionnaglaze@google.com>