Make test KDS tri-value, deprecate --test_use_kds.
While it's useful to provide locally cached certificates for a small
pool of test machines, and to test more generally with KDS, we'd like to
be able to test whether VCEK caching is working correctly. For this, we
explicitly request --test_kds=none, and not just hope that the local
cache doesn't hit.

The --test_use_kds=true value is --test_kds=amd. The false value is
"cache".

Signed-off-by: Dionna Glaze <dionnaglaze@google.com>