Add CertTableOptions for "extra" certs
Deprecate firmware_cert in favor of an Extras map that can contain any
number of extra certificates that can be validated with provided
options.

This adds to the basic "policy engine" that is validation primarily to
have a simple entry level user experience. More sophisticated policy
engines are not planned for this project.

Signed-off-by: Dionna Glaze <dionnaglaze@google.com>