Fix check tool product behavior on hardware
The product and stepping values that are default may not be the same
across all testing environments, so add product_name as a flag to
testing library. The stepping value does not necessarily need to be
checked, so deprecate the stepping field of SevProduct and replace with
a UInt32Value that can be nil.

The checkProductName drops checking Stepping in favor of checking
the nullable MachineStepping field.

Signed-off-by: Dionna Glaze <dionnaglaze@google.com>