Deprecate Product/ProductString for ProductLine
There is inconsistent wording for product string arguments that leads to
some confusing input/expectation mismatches. This change not only fixes
some wording, but also allows tests that expect to validate a report's
product to only care about the product line, and not the product name,
which can include the stepping version.

The test condition relaxation allows a Milan fleet of mixed staging
values to pass verify_test.