GetCrlAndCheckRoot: always verify CRL
Before, when an CRL was set in the TrustedRoot, it's CRL was not checked.
Since the CRL is only checked on retrieval, setting an CRL disabled the CRL
verification. So either the user needs to check the CRL manually before
setting it or let the library request the CRL again. The second options
breaks attestation when being strict about CRL verification and the KDS
is not available.

Change this to always verify the CRL when calling SnpAttestation.

Signed-off-by: Leonard Cohnen <lc@edgeless.systems>