abi: support Siena product line
This adds support for the "Siena" product line, which, as per the VCEK
and KDS interface spec (rev 1.0), is widely identical to the Genoa
product line. Furthermore, it aligns the parsing of the model line with
the specification by only referring to the extended model ID for
deriving the product name from the CPUID return value, which is what
Table 4 of the aforementioned specification suggests.